fix: xcodegen crashing if pods project not present already

+++ Example/scripts/post_gen.sh
@@ -1,15 +1,26 @@
 #!/bin/sh
 
-echo "✅ Podfile.lock removed"
-rm -f Podfile.lock
+SCHEMES="\n  - path: xcodegen/schemes.yml
+    relativePaths: NO"
 
-echo "✅ Pods removed"
-rm -rf Pods
+FOUND=`fgrep -c "schemes.yml" project.yml`
 
-echo "✅ xcworkspace removed"
-rm -rf *.xcworkspace
+if [ $FOUND -eq 1 ]; then
+  git checkout project.yml
+  exit 0
+else
+  echo "✅ Podfile.lock removed"
+  rm -f Podfile.lock
 
-if [ -f "Podfile" ]; then
-    echo "🚀 Installing Pods"
-    pod install
+  echo "✅ Pods removed"
+  rm -rf Pods
+
+  echo "✅ xcworkspace removed"
+  rm -rf *.xcworkspace
+
+  echo "🚀 Installing Pods"
+  pod install
+
+  echo "$SCHEMES" >> project.yml
+  xcodegen --quiet
 fi
